restructured files, fixed spacing, and added get all students path
