+++ api/models/User.py
@@ -30,10 +30,10 @@
 
 from mongoengine import *
 
-from api.models.general.LearningStyle import LearningStyle
-from api.models.user.LoginRecord import LoginRecord
-from api.models.user.SecurityQuestion import SecurityQuestion
-from api.models.user.Status import Status
+from api.models.generalEmbeds.LearningStyle import LearningStyle
+from api.models.userEmbeds.LoginRecord import LoginRecord
+from api.models.userEmbeds.SecurityQuestion import SecurityQuestion
+from api.models.userEmbeds.Status import Status
 
 
 class User(Document):
@@ -66,5 +66,3 @@ class User(Document):
     uuid = UUIDField(binary=True)
     _class = StringField()
     meta = {'collection': 'users'}
-
-
