+++ api/models/resourceEmbeds/CopyrightInfo.py
