+++ api/models/resourceEmbeds/Tags.py
