+++ api/services/utils/parsePutBody.py
@@ -1,5 +1,5 @@
 #  Aristotle Learning Platform: Luscinia Enterprises Assn.
-#  Copyright (C) 2020
+#  Copyright (C) 2021
 #      1261612 B.C. LTD.
 #
 #  This program is free software: you can redistribute it and/or modify
@@ -15,16 +15,16 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from mongoengine import *
+import uuid
+
+from dateutil.parser import isoparse
 
 
-class Course(Document):
-    name = StringField()
-    gradeLevel = IntField()
-    curriculum = StringField()
-    subject = StringField()
-    teacher = ObjectIdField()
-    users = ListField(ObjectIdField())
-    lessons = ListField(ObjectIdField())
-    created = DateField()
-    expires = DateField()
+def parsePutBody(body):
+    # manually parse dates and uuid if needed
+    if "copyrightInfo" in body and "publication" in body["copyrightInfo"]:
+        body["copyrightInfo"]["publication"] = isoparse(body["copyrightInfo"]["publication"]["$date"])
+    if "copyrightInfo" in body and "update" in body["copyrightInfo"]:
+        body["copyrightInfo"]["update"] = isoparse(body["copyrightInfo"]["update"]["$date"])
+    if "uuid" in body:
+        body["uuid"] = uuid.UUID(body["uuid"]["$uuid"])
